\docType{package}
\name{PROFANCY-package}
\alias{PROFANCY-package}
\alias{PROFANCY}
\title{The PROFANCY package}
\description{PROFANCY is a R platform to prioritize candidate disease metabolites by PROFANCY method which extend random walk analysis in the metabolic network with functional pathway nodes}
\section{Introduction}{
The main goal of PROFANCY is to prioritize candidate disease metabolites by PROFANCY method. This system provides the rank of candidate diseases metabolites which coud be provided by users or this system.  

}

\author{ Qianlan Yao <yaoqianlan@yahoo.com> Desi Shang <sds_46@163.com> Chunquan Li <lcqbio@aliyun.com>}
