\name{envData}
\alias{envData}
\alias{DiseaseInfList}
\alias{EHMNAddPathInfNetwork}
\alias{KEGGAddPathInfNetwork}
\alias{MetaboliteInf}
\alias{ProstateCandidates}
\alias{ProstateSeeds}
\docType{data}

\title{The variables in the environment variable envData of the system}
\description{
  The variables in the environment variable envData of the system.
}
\details{
The environment variable includes the variable \code{DiseaseInfList}, \code{EHMNAddPathInfNetwork}, \code{KEGGAddPathInfNetwork}, \code{MetaboliteInf}, \code{ProstateCandidates}, \code{ProstateSeeds}, etc.

}
\format{An environment variable}
\author{ Qianlan Yao <yaoqianlan@yahoo.com> Desi Shang <sds_46@163.com> Chunquan Li <lcqbio@aliyun.com>}
\seealso{\code{\link{initializeData}}}
\keyword{file}