\name{PROFANCY-internal}
\alias{RandomWalk2igraph}
\alias{rw}
\alias{getGraph}
\alias{getPriList}
\title{PROFANCY internal functions}
\description{
  These are functions internally called by other functions in the package \code{PROFANCY} and not meant to be called by the user.
}

\usage{
RandomWalk2igraph(igraphM,VertexWeight,EdgeWeight=TRUE,gamma=0.7)
rw(W,p0,gamma)
getGraph(network)
getPriList(Graph,seed)

}

\author{ Qianlan Yao <yaoqianlan@yahoo.com> Desi Shang <sds_46@163.com> Chunquan Li <lcqbio@aliyun.com>}

\keyword{internal}
